pyproject.toml
[tool.poetry]
name = "deep-dark-analyzer"
version = "0.1.0"
description = "Analyzer for press release for CBR"
authors = ["404"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
pydantic-settings = "^2.0.3"
SQLAlchemy = "^2.0.20"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"
SQLAlchemy-Utils = "^0.41.1"
uvicorn = "^0.23.2"
loguru = "^0.7.1"
gunicorn = "^21.2.0"
Jinja2 = "^3.1.2"
celery = "^5.3.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
mypy = "^1.5.1"
black = "^23.9.0"
flake8 = "^6.1.0"
bandit = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py311"]

[tool.isort]
known_local_folder = "backend"
py_version = "311"
profile = "black"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "sqlalchemy_utils.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true